refact(aws): renaming and file structures
