+++ src/providers/cloud-providers/aws/interfaces/NetworkService.ts
@@ -1,0 +1,8 @@
+import { Region } from "../../../../core/domain";
+
+export interface NetworkService {
+    /**
+     * Retrieves the public IP address from an EC2 instance
+     */
+    getPublicIp(instanceId: string, region: Region): Promise<string>;
+}
