+++ src/providers/cloud-providers/aws/interfaces/SecurityGroupService.ts
@@ -1,0 +1,13 @@
+import { Region } from "../../../../core/domain";
+
+export interface SecurityGroupService {
+    /**
+     * Creates a security group for the TF2 server with necessary rules
+     */
+    create(serverId: string, region: Region): Promise<string>;
+
+    /**
+     * Deletes the security group for the TF2 server
+     */
+    delete(serverId: string, region: Region): Promise<void>;
+}
