+++ src/providers/cloud-providers/aws/services/index.ts
@@ -1,5 +1,5 @@
-export * from './EC2InstanceService';
-export * from './ECSServiceManager';
-export * from './TaskDefinitionService';
-export * from './SecurityGroupService';
-export * from './NetworkService';
+export * from './DefaultEC2InstanceService';
+export * from './DefaultECSServiceManager';
+export * from './DefaultTaskDefinitionService';
+export * from './DefaultSecurityGroupService';
+export * from './DefaultNetworkService';
